requirements.txt
# Core ML / NLP
torch==2.2.2
transformers==4.38.2
sentence-transformers==2.7.0
tokenizers==0.15.2
huggingface-hub==0.25.0

# Vector store
faiss-cpu==1.8.0

# LLM
groq>=0.5.0

# Document processing
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
pygments>=2.17.0
gitpython>=3.1.40
tiktoken>=0.6.0

# RAG / Retrieval
rank-bm25>=0.2.2
nltk>=3.8.1

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Evaluation
ragas>=0.1.7
rouge-score>=0.1.2

# Experiment tracking
mlflow>=2.11.0

# UI
streamlit>=1.32.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
rich>=13.7.0
httpx>=0.27.0
